requirements.txt
pyttsx3>="2.90"
speechrecognition>="3.10.0"
openai-whisper>="20230314"
pyaudio>="0.2.13"
openai==1.8.0
# openai==0.28
pysoundfile==0.9.0.post1
python-dotenv==1.0.0
nltk==3.8.1
inject>=5.2.0
llama_cpp_python==0.2.31